+++ src/test/scala/dxWDL/dx/DxdaManifestTest.scala
@@ -5,6 +5,8 @@ import java.nio.file.{Path, Paths}
 import org.scalatest.{FlatSpec, Matchers}
 import spray.json._
 
+import dxWDL.base.Utils
+
 class DxdaManifestTest extends FlatSpec with Matchers {
 
     val TEST_PROJECT = "dxWDL_playground"
@@ -17,10 +19,42 @@ class DxdaManifestTest extends FlatSpec with Matchers {
                                         |the platform on staging.""".stripMargin)
         }
 
+    // convert :
+    //  { "1": {"md5":"71565d7f4dc0760457eb252a31d45964","size":42,"state":"complete"}}
+    // to
+    //  { "1": {"md5":"71565d7f4dc0760457eb252a31d45964","size":42}}
+    //
+    private def cleanPart(jsv: JsValue) : JsValue = {
+        val fields = jsv.asJsObject.fields.filter{ case (k, _) =>
+            k != "state"
+        }.toMap
+        JsObject(fields)
+    }
+
+    // describe a platform file, including its parts, with the dx-toolkit. We can then compare
+    // it to what we get from our DxdaManifest code.
+    //
+    private def describeWithParts(dxFile: DXFile) : JsValue = {
+        val cmd = s"""dx api ${dxFile.getId} describe '{"fields": { "parts" : true } }'"""
+        val (stdout, stderr) = Utils.execCommand(cmd)
+        val jsv = stdout.parseJson
+
+        // remove the "state" field from all the parts
+        val fields = jsv.asJsObject.fields.map{
+            case ("parts", parts) =>
+                "parts" -> JsObject(parts.asJsObject.fields.map{
+                                        case (k, v) => k -> cleanPart(v)
+                                    }.toMap)
+            case (k, v) =>
+                k -> v
+        }.toMap
+        JsObject(fields)
+    }
+
     it should "create manifests for dxda" in {
 
         val fileDir : Map[String, Path] = Map(
-            s"dx://${TEST_PROJECT}:/test_data/fileA" -> Paths.get("inputs/A"),
+            s"dx://${TEST_PROJECT}:/test_data/fileA" -> Paths.get("inputs/A") ,
             s"dx://${TEST_PROJECT}:/test_data/fileB" -> Paths.get("inputs/B"),
             s"dx://${TEST_PROJECT}:/test_data/fileC" -> Paths.get("inputs/C")
         )
@@ -38,36 +72,25 @@ class DxdaManifestTest extends FlatSpec with Matchers {
         // create a manifest
         val manifest : DxdaManifest = DxdaManifest.apply(filesInManifest)
 
-        val fileA = JsObject("id" -> JsString("file-FGqFGBQ0ffPPkYP19gBvFkZy"),
-                             "name" -> JsString("A"),
-                             "folder" -> JsString("inputs"),
-                             "parts" -> JsObject(
-                                 "1" -> JsObject(
-                                     "size" -> JsNumber(42),
-                                     "md5" -> JsString("71565d7f4dc0760457eb252a31d45964")
-                                 )))
-        val fileB = JsObject("id" -> JsString("file-FGqFJ8Q0ffPGVz3zGy4FK02P"),
-                             "name" -> JsString("B"),
-                             "folder" -> JsString("inputs"),
-                             "parts" -> JsObject(
-                                 "1" -> JsObject(
-                                     "size" -> JsNumber(8632),
-                                     "md5" -> JsString("903fb36e76d77c1ac95e90dbde9508df")
-                                 )))
-        val fileC = JsObject("id" -> JsString("file-FGzzpkQ0ffPJX74548Vp6670"),
-                             "name" -> JsString("C"),
-                             "folder" -> JsString("inputs"),
-                             "parts" -> JsObject(
-                                 "1" -> JsObject(
-                                     "size" -> JsNumber(42),
-                                     "md5" -> JsString("57a8ec78ad9b08b2add5b6656017e34c")
-                                 )))
+
+        // compare to data obtained with dx-toolkit
+        val expected : Vector[JsValue] = resolvedObjects.map{
+            case (dxPath, dataObj) =>
+                val dxFile = dataObj.asInstanceOf[DXFile]
+                val local : Path = fileDir(dxPath)
+                val jsv = describeWithParts(dxFile)
+
+                // add the target folder and name
+                val fields = jsv.asJsObject.fields ++  Map(
+                    "name" -> JsString(local.toFile().getName()),
+                    "folder" -> JsString(local.toFile().getParent())
+                )
+                JsObject(fields)
+        }.toVector
 
         manifest shouldBe(
             DxdaManifest(
-                JsObject(
-                    dxTestProject.getId -> JsArray(Vector(fileA, fileB, fileC)))
-            )
-        )
+                JsObject(dxTestProject.getId -> JsArray(expected))
+            ))
     }
 }